#79 - refactor: auth 패키지 리팩토링
