+++ src/main/java/com/gxdxx/instagram/domain/reply/application/ReplyDeleteService.java
@@ -3,7 +3,7 @@ package com.gxdxx.instagram.domain.reply.application;
 import com.gxdxx.instagram.global.common.dto.response.SuccessResponse;
 import com.gxdxx.instagram.domain.reply.domain.Reply;
 import com.gxdxx.instagram.domain.reply.exception.ReplyNotFoundException;
-import com.gxdxx.instagram.global.auth.UnauthorizedAccessException;
+import com.gxdxx.instagram.global.auth.exception.UnauthorizedAccessException;
 import com.gxdxx.instagram.domain.reply.dao.ReplyRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
