+++ src/main/java/com/gxdxx/instagram/domain/reply/application/ReplyUpdateService.java
@@ -4,7 +4,7 @@ import com.gxdxx.instagram.domain.reply.dto.request.ReplyUpdateRequest;
 import com.gxdxx.instagram.domain.reply.dto.response.ReplyUpdateResponse;
 import com.gxdxx.instagram.domain.reply.domain.Reply;
 import com.gxdxx.instagram.domain.reply.exception.ReplyNotFoundException;
-import com.gxdxx.instagram.global.auth.UnauthorizedAccessException;
+import com.gxdxx.instagram.global.auth.exception.UnauthorizedAccessException;
 import com.gxdxx.instagram.domain.reply.dao.ReplyRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
