+++ src/main/java/com/gxdxx/instagram/domain/user/application/UserLogoutService.java
@@ -3,7 +3,7 @@ package com.gxdxx.instagram.domain.user.application;
 import com.gxdxx.instagram.global.config.jwt.TokenProvider;
 import com.gxdxx.instagram.global.common.dto.response.SuccessResponse;
 import com.gxdxx.instagram.domain.user.domain.User;
-import com.gxdxx.instagram.global.auth.RefreshTokenInvalidException;
+import com.gxdxx.instagram.global.auth.exception.RefreshTokenInvalidException;
 import com.gxdxx.instagram.domain.user.exception.UserNotFoundException;
 import com.gxdxx.instagram.domain.user.dao.UserRepository;
 import com.gxdxx.instagram.global.redis.RedisService;
