+++ src/main/java/com/gxdxx/instagram/global/auth/application/AccessTokenCreateService.java
@@ -1,5 +1,7 @@
-package com.gxdxx.instagram.global.auth;
+package com.gxdxx.instagram.global.auth.application;
 
+import com.gxdxx.instagram.global.auth.exception.RefreshTokenInvalidException;
+import com.gxdxx.instagram.global.auth.dto.AccessTokenCreateRequest;
 import com.gxdxx.instagram.global.config.jwt.TokenProvider;
 import com.gxdxx.instagram.global.common.dto.response.SuccessResponse;
 import com.gxdxx.instagram.domain.user.domain.User;
