+++ src/main/java/com/gxdxx/instagram/global/auth/dto/AccessTokenCreateRequest.java
@@ -1,4 +1,4 @@
-package com.gxdxx.instagram.global.auth;
+package com.gxdxx.instagram.global.auth.dto;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import io.swagger.v3.oas.annotations.media.Schema;
