+++ src/main/java/com/gxdxx/instagram/global/auth/exception/RefreshTokenInvalidException.java
@@ -1,4 +1,4 @@
-package com.gxdxx.instagram.global.auth;
+package com.gxdxx.instagram.global.auth.exception;
 
 public class RefreshTokenInvalidException extends RuntimeException {
 }
