+++ src/main/java/com/gxdxx/instagram/global/auth/exception/UnauthorizedAccessException.java
@@ -1,4 +1,4 @@
-package com.gxdxx.instagram.global.auth;
+package com.gxdxx.instagram.global.auth.exception;
 
 public class UnauthorizedAccessException extends RuntimeException {
 }
