+++ src/main/java/com/gxdxx/instagram/global/error/advice/RestControllerExceptionAdvice.java
@@ -10,8 +10,8 @@ import com.gxdxx.instagram.domain.user.exception.NicknameAlreadyExistsException;
 import com.gxdxx.instagram.domain.user.exception.PasswordNotMatchException;
 import com.gxdxx.instagram.domain.user.exception.UserNotFoundException;
 import com.gxdxx.instagram.global.common.dto.response.ErrorResponse;
-import com.gxdxx.instagram.global.auth.RefreshTokenInvalidException;
-import com.gxdxx.instagram.global.auth.UnauthorizedAccessException;
+import com.gxdxx.instagram.global.auth.exception.RefreshTokenInvalidException;
+import com.gxdxx.instagram.global.auth.exception.UnauthorizedAccessException;
 import com.gxdxx.instagram.global.error.FileProcessingException;
 import com.gxdxx.instagram.global.error.InvalidRequestException;
 import jakarta.validation.ConstraintViolationException;
